+++ src/components/shared/Search/Search.tsx
@@ -1,28 +1,12 @@
 import * as React from 'react';
-import * as Autosuggest from 'react-autosuggest';
+import AutoComplete from 'material-ui/AutoComplete';
 
-import './Search.css';
-
-const theme = {
-  input: 'autosuggest_input',
-  inputFocused: 'autosuggest_input--focused',
-  suggestionsList: 'autosuggest_suggestions-list',
-  suggestion: 'autosuggest_suggestion',
-  suggestionHighlighted: 'autosuggest_suggestion--highlighted'
-};
-
-const renderSuggestion = (suggestion: string) => (
-  <div>
-    {suggestion}
-  </div>
-);
-
-class Search extends React.Component<ISearchProps, { value: string, suggestions: Array<string> }> {
+class Search extends React.Component<ISearchProps, { searchText: string, suggestions: Array<string> }> {
   constructor(props: ISearchProps) {
     super(props);
 
     this.state = {
-      value: '',
+      searchText: '',
       suggestions: props.suggestions
     };
   }
@@ -33,48 +17,37 @@ class Search extends React.Component<ISearchProps, { value: string, suggestions:
     });
   }
 
-  onChange = (event: any, { newValue }: { newValue: string }) => {
+  handleUpdateInput = (searchText: string) => {
+    if (searchText === '') {
+      this.props.onClear();
+    }
+    this.props.onChange(searchText);
     this.setState({
-      value: newValue
+      searchText: searchText
     });
   }
 
-  onSuggestionsFetchRequested = ({ value }: { value: string }) => {
-    this.props.onChange(value);
-    this.setState({
-      suggestions: this.props.suggestions
-    });
-  }
-
-  onSuggestionsClearRequested = () => {
+  handleNewRequest = () => {
+    this.props.onSelected(this.state.searchText);
     this.props.onClear();
     this.setState({
-      suggestions: []
+      searchText: '',
     });
   }
 
   render() {
-    const { value, suggestions } = this.state;
-
-    const inputProps = {
-      placeholder: this.props.placeholder,
-      value,
-      onChange: this.onChange
-    };
-
     return (
       <div>
-        <Autosuggest
-          inputProps={inputProps}
-          suggestions={suggestions}
-          getSuggestionValue={(suggestion: string) => suggestion}
-          renderSuggestion={renderSuggestion}
-          theme={theme}
-          onSuggestionsFetchRequested={this.onSuggestionsFetchRequested}
-          onSuggestionsClearRequested={this.onSuggestionsClearRequested}
-          onSuggestionSelected={(e: any, { suggestionValue }: { suggestionValue: string }) => {
-            this.props.onSelected(suggestionValue);
-          }}
+        <AutoComplete
+          hintText={this.props.placeholder}
+          searchText={this.state.searchText}
+          onUpdateInput={this.handleUpdateInput}
+          onNewRequest={this.handleNewRequest}
+          dataSource={this.state.suggestions}
+          filter={(searchText, key) => (key.indexOf(searchText.toLowerCase()) !== -1)}
+          openOnFocus={true}
+          maxSearchResults={10}
+          fullWidth={true}
         />
       </div>
     );
